+++ Views/Albums/Details.cshtml
@@ -33,6 +33,6 @@
 <ul>
     @foreach (var song in Model.Songs)
     {
-        <li>@song.TrackSong: "@song.Title</li>
+        <li>@song.TrackSong "@song.Title"</li>
     }
 </ul>
