--- Views/PlaylistSongs/Create.cshtml
@@ -1,37 +1,0 @@
-﻿@model CodeFirstMusicSystem.Models.PlaylistSong
-
-@{
-    ViewData["Title"] = "Create";
-    Layout = "~/Views/Shared/_Layout.cshtml";
-}
-
-<h1>Create</h1>
-
-<h4>PlaylistSong</h4>
-<hr />
-<div class="row">
-    <div class="col-md-4">
-        <form asp-action="Create">
-            <div asp-validation-summary="ModelOnly" class="text-danger"></div>
-            <div class="form-group">
-                <label asp-for="PlaylistId" class="control-label"></label>
-                <select asp-for="PlaylistId" class ="form-control" asp-items="ViewBag.PlaylistId"></select>
-            </div>
-            <div class="form-group">
-                <label asp-for="SongId" class="control-label"></label>
-                <select asp-for="SongId" class ="form-control" asp-items="ViewBag.SongId"></select>
-            </div>
-            <div class="form-group">
-                <input type="submit" value="Create" class="btn btn-primary" />
-            </div>
-        </form>
-    </div>
-</div>
-
-<div>
-    <a asp-action="Index">Back to List</a>
-</div>
-
-@section Scripts {
-    @{await Html.RenderPartialAsync("_ValidationScriptsPartial");}
-}
