--- Views/PlaylistSongs/Delete.cshtml
@@ -1,34 +1,0 @@
-﻿@model CodeFirstMusicSystem.Models.PlaylistSong
-
-@{
-    ViewData["Title"] = "Delete";
-    Layout = "~/Views/Shared/_Layout.cshtml";
-}
-
-<h1>Delete</h1>
-
-<h3>Are you sure you want to delete this?</h3>
-<div>
-    <h4>PlaylistSong</h4>
-    <hr />
-    <dl class="row">
-        <dt class = "col-sm-2">
-            @Html.DisplayNameFor(model => model.Playlist)
-        </dt>
-        <dd class = "col-sm-10">
-            @Html.DisplayFor(model => model.Playlist.Id)
-        </dd>
-        <dt class = "col-sm-2">
-            @Html.DisplayNameFor(model => model.Song)
-        </dt>
-        <dd class = "col-sm-10">
-            @Html.DisplayFor(model => model.Song.Id)
-        </dd>
-    </dl>
-    
-    <form asp-action="Delete">
-        <input type="hidden" asp-for="Id" />
-        <input type="submit" value="Delete" class="btn btn-danger" /> |
-        <a asp-action="Index">Back to List</a>
-    </form>
-</div>
