--- Views/PlaylistSongs/Details.cshtml
@@ -1,31 +1,0 @@
-﻿@model CodeFirstMusicSystem.Models.PlaylistSong
-
-@{
-    ViewData["Title"] = "Details";
-    Layout = "~/Views/Shared/_Layout.cshtml";
-}
-
-<h1>Details</h1>
-
-<div>
-    <h4>PlaylistSong</h4>
-    <hr />
-    <dl class="row">
-        <dt class = "col-sm-2">
-            @Html.DisplayNameFor(model => model.Playlist)
-        </dt>
-        <dd class = "col-sm-10">
-            @Html.DisplayFor(model => model.Playlist.Id)
-        </dd>
-        <dt class = "col-sm-2">
-            @Html.DisplayNameFor(model => model.Song)
-        </dt>
-        <dd class = "col-sm-10">
-            @Html.DisplayFor(model => model.Song.Id)
-        </dd>
-    </dl>
-</div>
-<div>
-    <a asp-action="Edit" asp-route-id="@Model?.Id">Edit</a> |
-    <a asp-action="Index">Back to List</a>
-</div>
